Use caching in device
